don't crash if multiplayer but old client on new enough server.
e.g. a 0.4.17[.x] server is enough to run everything normally, but a
0.4.16 or older client doesn't call on_rightclick if the node also has a
formspec set, which leaves minislots.player_last_machine_def unset/nil and
crashes the server.

If the server or singleplayer is 0.4.16 or older, a crash expected due to
newer functions that are used; this cannot be fixed (and it doesn't need to
be anyway :-P ).